Update login.py
I merged the function that's called when you click register, to first check whether the user is on the login screen, or the register screen, and if it's on the registration screen already, add the inputted details into the database. However, if this was a real application, there would need to be some feedback for the user to know something actually happened. Text simply disappearing off the screen would have this effect, but better yet, text thanking the user for registering their details.